refactor(mock): transit comments into separate mock file
